refactor(app): changes all file to src folder
