+++ backend/package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "@prisma/client": "^4.16.2",
     "dotenv": "^16.3.1",
+    "envalid": "^7.3.1",
     "express": "^4.18.2"
   },
   "devDependencies": {
