+++ backend/src/app.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { PORT } from "../config";
+import { PORT } from "./config";
 export class App {
   public app: express.Application;
   public port: string | number;
