+++ backend/src/config/index.ts
@@ -2,4 +2,4 @@ import { config } from "dotenv";
 
 config({ path: ".env" });
 
-export const { PORT } = process.env;
+export const { PORT, LOG_DIR } = process.env;
