+++ backend/src/index.ts
@@ -1,4 +1,7 @@
 import { App } from "./app";
+import { validateEnv } from "./utils/validateEnv";
+
+validateEnv();
 
 const app = new App();
