+++ backend/src/utils/validateEnv.ts
@@ -1,0 +1,8 @@
+import { cleanEnv, port, str } from "envalid";
+
+export const validateEnv = () => {
+  cleanEnv(process.env, {
+    PORT: port(),
+    LOG_DIR: str(),
+  });
+};
